[REFACTOR]: 이미지 파일 구조 변환
