+++ src/main/java/the_t/mainproject/domain/image/dto/ImageRes.java
@@ -1,4 +1,4 @@
-package the_t.mainproject.domain.post.dto.res;
+package the_t.mainproject.domain.image.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
