+++ src/main/java/the_t/mainproject/domain/post/application/PostService.java
@@ -2,7 +2,7 @@ package the_t.mainproject.domain.post.application;
 
 import org.springframework.web.multipart.MultipartFile;
 import the_t.mainproject.domain.post.dto.req.PostReq;
-import the_t.mainproject.domain.post.dto.res.ImageRes;
+import the_t.mainproject.domain.image.dto.ImageRes;
 import the_t.mainproject.domain.post.dto.res.LikedCountRes;
 import the_t.mainproject.domain.post.dto.res.PostDetailRes;
 import the_t.mainproject.domain.post.dto.res.PostListRes;
