+++ src/main/java/the_t/mainproject/domain/post/application/PostServiceImpl.java
@@ -15,13 +15,13 @@ import the_t.mainproject.domain.liked.domain.Liked;
 import the_t.mainproject.domain.liked.domain.repository.LikedRepository;
 import the_t.mainproject.domain.member.domain.Member;
 import the_t.mainproject.domain.member.domain.repository.MemberRepository;
-import the_t.mainproject.domain.post.domain.Image;
+import the_t.mainproject.domain.image.domain.Image;
 import the_t.mainproject.domain.post.domain.Post;
 import the_t.mainproject.domain.post.domain.VoiceType;
-import the_t.mainproject.domain.post.domain.repository.ImageRepository;
+import the_t.mainproject.domain.image.domain.repository.ImageRepository;
 import the_t.mainproject.domain.post.domain.repository.PostRepository;
 import the_t.mainproject.domain.post.dto.req.PostReq;
-import the_t.mainproject.domain.post.dto.res.ImageRes;
+import the_t.mainproject.domain.image.dto.ImageRes;
 import the_t.mainproject.domain.post.dto.res.LikedCountRes;
 import the_t.mainproject.domain.post.dto.res.PostDetailRes;
 import the_t.mainproject.domain.post.dto.res.PostListRes;
